Fix changes and mega refactor incoming
