+++ teach/src/main/java/com/u/teach/controller/BaseController.java
@@ -1,10 +1,8 @@
 package com.u.teach.controller;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
+import android.view.ViewGroup;
+import com.bluelinelabs.conductor.RouterTransaction;
+import com.bluelinelabs.conductor.changehandler.FadeChangeHandler;
 import com.bluelinelabs.conductor.rxlifecycle.RxController;
 import com.u.teach.R;
 
@@ -13,29 +11,13 @@ import com.u.teach.R;
  */
 public abstract class BaseController extends RxController {
 
-    @Override
-    protected void onAttach(@NonNull final View view) {
-        super.onAttach(view);
-
-        if (getActionBar() != null) {
-            if (hasActionBar()) {
-                getActionBar().setTitle(" " + title());
-                getActionBar().setIcon(R.mipmap.ic_launcher);
-                getActionBar().show();
-            } else {
-                getActionBar().hide();
-            }
-        }
+    @SuppressWarnings("ConstantConditions")
+    protected void showDialog(BaseController controller) {
+        getChildRouter((ViewGroup) getActivity().findViewById(R.id.activity_dialog_container))
+            .setPopsLastView(true)
+            .setRoot(RouterTransaction.with(controller)
+                .popChangeHandler(new FadeChangeHandler())
+                .pushChangeHandler(new FadeChangeHandler()));
     }
-
-    protected boolean hasActionBar() {
-        return true;
-    }
-
-    protected @Nullable ActionBar getActionBar() {
-        return getActivity() == null ? null : ((AppCompatActivity) getActivity()).getSupportActionBar();
-    }
-
-    protected abstract @Nullable String title();
 
 }
